+++ backend/src/main/java/com/optimiza/clickbarber/model/Cliente.java
@@ -4,7 +4,6 @@ import jakarta.persistence.*;
 import lombok.*;
 
 import java.time.LocalDate;
-import java.util.UUID;
 
 @Table(name = "clientes")
 @Entity
@@ -19,7 +18,7 @@ public class Cliente {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String cpf;
     private String nome;
