+++ backend/src/main/java/com/optimiza/clickbarber/repository/BarbeiroRepository.java
@@ -1,7 +1,6 @@
 package com.optimiza.clickbarber.repository;
 
 import com.optimiza.clickbarber.model.Barbeiro;
-import com.optimiza.clickbarber.model.Cliente;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -10,7 +9,7 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Repository
-public interface BarbeiroRepository extends JpaRepository<Barbeiro, Integer> {
+public interface BarbeiroRepository extends JpaRepository<Barbeiro, Long> {
 
     List<Barbeiro> findByBarbeariaId(Integer id);
