+++ backend/src/main/java/com/optimiza/clickbarber/repository/ClienteRepository.java
@@ -2,13 +2,11 @@ package com.optimiza.clickbarber.repository;
 
 import com.optimiza.clickbarber.model.Cliente;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 
 import java.util.Optional;
 import java.util.UUID;
 
-public interface ClienteRepository extends JpaRepository<Cliente, Integer> {
+public interface ClienteRepository extends JpaRepository<Cliente, Long> {
 
     Optional<Cliente> findByUsuarioId(UUID usuarioId);
