+++ backend/src/main/resources/db/migration/V8__cria_tabela_agendamentos_barbeiros.sql
@@ -1,6 +1,6 @@
 CREATE TABLE agendamentos_barbeiros (
     agendamento_id UUID NOT NULL,
-    barbeiro_id INTEGER NOT NULL,
+    barbeiro_id BIGINT NOT NULL,
     PRIMARY KEY (agendamento_id, barbeiro_id),
     FOREIGN KEY (agendamento_id) REFERENCES agendamentos(id),
     FOREIGN KEY (barbeiro_id) REFERENCES barbeiros(id)
